chore: generalize tab interface
