+++ src/components/Chip/BetaChip.tsx
@@ -1,0 +1,10 @@
+import React from "react";
+import { Chip } from "@site/src/components/Chip/Chip";
+
+export function BetaChip() {
+  return (
+    <Chip size={"small"} shape={"rounded"}>
+      Beta
+    </Chip>
+  );
+}
