--- src/components/Tabs/ICTabItem.tsx
@@ -1,6 +1,0 @@
-import React from "react";
-import TabItem, { Props as TabItemProps } from "@theme/TabItem";
-
-export function ICTabItem(props: TabItemProps) {
-  return <TabItem {...props} />;
-}
